Add files via upload
Made changes in both RollDice and TheFoodBot to make the experience of entering in choices more pleasant for the user. Additionally, changes were made to the help menu, and a list of restaurant types the user is restricted to has been included. If the user searches for a type of restaurant that is not in that list, they will be informed of the available choices and asked if they would like to search again.